@@ -5,7 +5,7 @@ build-backend = "poetry.core.masonry.api"
 [tool.poetry]
 name = "integrify"
 packages = [{include = "integrify", from="src"}]
-version = "2.0.1"
+version = "2.1.0"
 description = "Integrify API inteqrasiyalarını rahatlaşdıran sorğular kitabaxanasıdır."
 license = "GPL-3.0-or-later"
 authors = [
@@ -74,7 +74,7 @@ mkdocs-panzoom-plugin = "^0.1.3"
 griffe-pydantic = "^1.0.0"
 
 [tool.ruff]
-target-version = "py38"
+target-version = "py39"
 line-length = 100
 src = ["src", "tests"]
 
